requirements.txt
Django==5.1.3
python-dotenv>=1.0.0
social-auth-app-django>=5.4.0
django-allauth==0.59.0
django-otp==1.3.0
qrcode==7.4.2
pillow==10.2.0
django-two-factor-auth==1.15.5
phonenumbers==8.13.27
